Fixed issue related to int to double conversion

+++ lib/widgets/all_currencies_widget.dart
@@ -67,7 +67,7 @@ class _AllCurrenciesWidgetState extends State<AllCurrenciesWidget> {
           "${currency.currencyName}",
         ),
         trailing: Text(
-          (currency.currencyValue as double).toStringAsFixed(3),
+          currency.currencyValue.toStringAsFixed(3),
           style: TextStyle(
             fontSize: 20.0,
             fontWeight: FontWeight.bold
